# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Translations panel enpoints",
    "description": "Api endpoints for translations panel v1",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/language/": {
      "get": {
        "tags": ["Language data"],
        "summary": "Gets list of languages",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Languages"
            }
          }
        }
      },
      "post": {
        "tags": ["Language data"],
        "summary": "Creates new language",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "en-UK",
                  "description": "ISO code of language. Example en-US",
                  "enum": [
                    "af-ZA",
                    "am-ET",
                    "ar-AE",
                    "ar-BH",
                    "ar-DZ",
                    "ar-EG",
                    "ar-IQ",
                    "ar-JO",
                    "ar-KW",
                    "ar-LB",
                    "ar-LY",
                    "ar-MA",
                    "ar-OM",
                    "ar-QA",
                    "ar-SA",
                    "ar-SD",
                    "ar-SY",
                    "ar-TN",
                    "ar-YE",
                    "as-IN",
                    "az-Cyrl-AZ",
                    "az-Latn-AZ",
                    "be-BY",
                    "bg-BG",
                    "bn-BD",
                    "bn-IN",
                    "bs-Cyrl-BA",
                    "bs-Latn-BA",
                    "ca-ES",
                    "cs-CZ",
                    "cy-GB",
                    "da-DK",
                    "de-AT",
                    "de-CH",
                    "de-DE",
                    "de-LI",
                    "de-LU",
                    "dv-MV",
                    "el-GR",
                    "en-029",
                    "en-AU",
                    "en-BZ",
                    "en-CA",
                    "en-GB",
                    "en-IE",
                    "en-IN",
                    "en-JM",
                    "en-MY",
                    "en-NZ",
                    "en-PH",
                    "en-SG",
                    "en-TT",
                    "en-US",
                    "en-ZA",
                    "en-ZW",
                    "es-AR",
                    "es-BO",
                    "es-CL",
                    "es-CO",
                    "es-CR",
                    "es-DO",
                    "es-EC",
                    "es-ES",
                    "es-GT",
                    "es-HN",
                    "es-MX",
                    "es-NI",
                    "es-PA",
                    "es-PE",
                    "es-PR",
                    "es-PT",
                    "es-PY",
                    "es-SV",
                    "es-US",
                    "es-UY",
                    "es-VE",
                    "et-EE",
                    "eu-ES",
                    "fa-IR",
                    "fi-FI",
                    "fil-PH",
                    "fo-FO",
                    "fr-BE",
                    "fr-CA",
                    "fr-CH",
                    "fr-FR",
                    "fr-LU",
                    "fr-MC",
                    "fy-NL",
                    "ga-IE",
                    "gd-GB",
                    "gl-ES",
                    "gsw-FR",
                    "gu-IN",
                    "ha-Latn-NG",
                    "he-IL",
                    "hi-IN",
                    "hr-BA",
                    "hr-HR",
                    "hsb-DE",
                    "hu-HU",
                    "hy-AM",
                    "id-ID",
                    "ig-NG",
                    "ii-CN",
                    "is-IS",
                    "it-CH",
                    "it-IT",
                    "iu-Cans-CA",
                    "iu-Latn-CA",
                    "ja-JP",
                    "ka-GE",
                    "kk-KZ",
                    "kl-GL",
                    "km-KH",
                    "kn-IN",
                    "kok-IN",
                    "ko-KR",
                    "ky-KG",
                    "lb-LU",
                    "lo-LA",
                    "lt-LT",
                    "lv-LV",
                    "mi-NZ",
                    "mk-MK",
                    "ml-IN",
                    "mn-MN",
                    "mn-Mong-CN",
                    "moh-CA",
                    "mr-IN",
                    "ms-BN",
                    "ms-MY",
                    "mt-MT",
                    "nb-NO",
                    "ne-NP",
                    "nl-BE",
                    "nl-NL",
                    "nn-NO",
                    "nso-ZA",
                    "oc-FR",
                    "or-IN",
                    "pa-IN",
                    "pt-BR",
                    "pt-PT",
                    "qut-GT",
                    "quz-BO",
                    "quz-EC",
                    "quz-PE",
                    "rm-CH",
                    "ro-RO",
                    "ru-RU",
                    "rw-RW",
                    "sa-IN",
                    "sah-RU",
                    "se-FI",
                    "se-NO",
                    "se-SE",
                    "si-LK",
                    "sk-SK",
                    "sl-SI",
                    "sma-NO",
                    "sma-SE",
                    "smj-NO",
                    "smj-SE",
                    "smn-FI",
                    "sms-FI",
                    "sq-AL",
                    "sr-Cyrl-BA",
                    "sr-Cyrl-CS",
                    "sr-Cyrl-ME",
                    "sr-Cyrl-RS",
                    "sr-Latn-BA",
                    "sr-Latn-CS",
                    "sr-Latn-ME",
                    "sr-Latn-RS",
                    "sv-FI",
                    "sv-SE",
                    "sw-KE",
                    "syr-SY",
                    "ta-IN",
                    "te-IN",
                    "tg-Cyrl-TJ",
                    "th-TH",
                    "tk-TM",
                    "tn-ZA",
                    "tr-TR",
                    "tt-RU",
                    "tzm-Latn-DZ",
                    "ug-CN",
                    "uk-UA",
                    "ur-PK",
                    "uz-Cyrl-UZ",
                    "uz-Latn-UZ",
                    "vi-VN",
                    "wo-SN",
                    "xh-ZA",
                    "yo-NG",
                    "zh-CN",
                    "zh-HK",
                    "zh-MO",
                    "zh-SG",
                    "zh-TW",
                    "zu-ZA"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Language"
            }
          },
          "409": {
            "description": "DUPLICATE",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Value {\"name\":\"en-UK\"} already exists"
                }
              }
            }
          },
          "400": {
            "description": "Missing argument",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Missing name in the playload"
                }
              }
            }
          }
        }
      }
    },
    "/language/{language}": {
      "get": {
        "tags": ["Language data"],
        "summary": "Gets data for given language",
        "parameters": [
          {
            "in": "path",
            "name": "language",
            "description": "ISO code of language. Example en-US",
            "required": true,
            "type": "string",
            "enum": [
              "af-ZA",
              "am-ET",
              "ar-AE",
              "ar-BH",
              "ar-DZ",
              "ar-EG",
              "ar-IQ",
              "ar-JO",
              "ar-KW",
              "ar-LB",
              "ar-LY",
              "ar-MA",
              "ar-OM",
              "ar-QA",
              "ar-SA",
              "ar-SD",
              "ar-SY",
              "ar-TN",
              "ar-YE",
              "as-IN",
              "az-Cyrl-AZ",
              "az-Latn-AZ",
              "be-BY",
              "bg-BG",
              "bn-BD",
              "bn-IN",
              "bs-Cyrl-BA",
              "bs-Latn-BA",
              "ca-ES",
              "cs-CZ",
              "cy-GB",
              "da-DK",
              "de-AT",
              "de-CH",
              "de-DE",
              "de-LI",
              "de-LU",
              "dv-MV",
              "el-GR",
              "en-029",
              "en-AU",
              "en-BZ",
              "en-CA",
              "en-GB",
              "en-IE",
              "en-IN",
              "en-JM",
              "en-MY",
              "en-NZ",
              "en-PH",
              "en-SG",
              "en-TT",
              "en-US",
              "en-ZA",
              "en-ZW",
              "es-AR",
              "es-BO",
              "es-CL",
              "es-CO",
              "es-CR",
              "es-DO",
              "es-EC",
              "es-ES",
              "es-GT",
              "es-HN",
              "es-MX",
              "es-NI",
              "es-PA",
              "es-PE",
              "es-PR",
              "es-PT",
              "es-PY",
              "es-SV",
              "es-US",
              "es-UY",
              "es-VE",
              "et-EE",
              "eu-ES",
              "fa-IR",
              "fi-FI",
              "fil-PH",
              "fo-FO",
              "fr-BE",
              "fr-CA",
              "fr-CH",
              "fr-FR",
              "fr-LU",
              "fr-MC",
              "fy-NL",
              "ga-IE",
              "gd-GB",
              "gl-ES",
              "gsw-FR",
              "gu-IN",
              "ha-Latn-NG",
              "he-IL",
              "hi-IN",
              "hr-BA",
              "hr-HR",
              "hsb-DE",
              "hu-HU",
              "hy-AM",
              "id-ID",
              "ig-NG",
              "ii-CN",
              "is-IS",
              "it-CH",
              "it-IT",
              "iu-Cans-CA",
              "iu-Latn-CA",
              "ja-JP",
              "ka-GE",
              "kk-KZ",
              "kl-GL",
              "km-KH",
              "kn-IN",
              "kok-IN",
              "ko-KR",
              "ky-KG",
              "lb-LU",
              "lo-LA",
              "lt-LT",
              "lv-LV",
              "mi-NZ",
              "mk-MK",
              "ml-IN",
              "mn-MN",
              "mn-Mong-CN",
              "moh-CA",
              "mr-IN",
              "ms-BN",
              "ms-MY",
              "mt-MT",
              "nb-NO",
              "ne-NP",
              "nl-BE",
              "nl-NL",
              "nn-NO",
              "nso-ZA",
              "oc-FR",
              "or-IN",
              "pa-IN",
              "pt-BR",
              "pt-PT",
              "qut-GT",
              "quz-BO",
              "quz-EC",
              "quz-PE",
              "rm-CH",
              "ro-RO",
              "ru-RU",
              "rw-RW",
              "sa-IN",
              "sah-RU",
              "se-FI",
              "se-NO",
              "se-SE",
              "si-LK",
              "sk-SK",
              "sl-SI",
              "sma-NO",
              "sma-SE",
              "smj-NO",
              "smj-SE",
              "smn-FI",
              "sms-FI",
              "sq-AL",
              "sr-Cyrl-BA",
              "sr-Cyrl-CS",
              "sr-Cyrl-ME",
              "sr-Cyrl-RS",
              "sr-Latn-BA",
              "sr-Latn-CS",
              "sr-Latn-ME",
              "sr-Latn-RS",
              "sv-FI",
              "sv-SE",
              "sw-KE",
              "syr-SY",
              "ta-IN",
              "te-IN",
              "tg-Cyrl-TJ",
              "th-TH",
              "tk-TM",
              "tn-ZA",
              "tr-TR",
              "tt-RU",
              "tzm-Latn-DZ",
              "ug-CN",
              "uk-UA",
              "ur-PK",
              "uz-Cyrl-UZ",
              "uz-Latn-UZ",
              "vi-VN",
              "wo-SN",
              "xh-ZA",
              "yo-NG",
              "zh-CN",
              "zh-HK",
              "zh-MO",
              "zh-SG",
              "zh-TW",
              "zu-ZA"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Language"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Language en-ES not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Language": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1,
          "description": "Id of language entity"
        },
        "createdAt": {
          "type": "string",
          "example": 12215474745467,
          "description": "Unix time of last change"
        },
        "updateAt": {
          "type": "string",
          "example": 12215474745467,
          "description": "Unix time of last change"
        },
        "code": {
          "type": "string",
          "example": "en-US",
          "description": "ISO based language code"
        },
        "country": {
          "type": "string",
          "example": "United States of America",
          "description": "Country full name"
        },
        "isDisabled": {
          "type": "boolean",
          "example": false,
          "description": "Shows if language is in use"
        },
        "namespace": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": "1"
              },
              "createdAt": {
                "type": "string",
                "example": "2023-03-14T19:10:07.581Z"
              },
              "updateAt": {
                "type": "string",
                "example": "2023-03-14T19:10:07.581Z"
              },
              "name": {
                "type": "string",
                "example": "common"
              },
              "languageId": {
                "type": "number",
                "example": "16"
              }
            }
          }
        }
      },
      "$ref": "#/definitions/Language"
    },
...
    "Languages": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": "16"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-03-14T19:10:00.552Z"
          },
          "updateAt": {
            "type": "string",
            "example": "2023-03-14T19:10:00.552Z"
          },
          "name": {
            "type": "string",
            "example": "en-US"
          },
          "isDisabled": {
            "type": "boolean"
          },
          "namespace": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": "1"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-03-14T19:10:07.581Z"
                },
                "updateAt": {
                  "type": "string",
                  "example": "2023-03-14T19:10:07.581Z"
                },
                "name": {
                  "type": "string",
                  "example": "common"
                },
                "languageId": {
                  "type": "number",
                  "example": "16"
                }
              }
            }
          }
        }
      }
    },
...
  }
}